+++ packages/admin-ui/src/node/constants.ts
@@ -1,0 +1,13 @@
+export const ALIASED_PACKAGES = [
+  "react",
+  "react-dom",
+  "react-router-dom",
+  "react-dnd",
+  "react-dnd-html5-backend",
+  "react-select",
+  "react-helmet-async",
+  "@tanstack/react-query",
+  "@tanstack/react-table",
+  "@emotion/react",
+  "medusa-react",
+] as const