+++ packages/admin-ui/ui/src/domain/orders/details/index.tsx
@@ -31,8 +31,10 @@ import { useEffect, useMemo, useState } from "react"
 import { useHotkeys } from "react-hotkeys-hook"
 import Avatar from "../../../components/atoms/avatar"
 import BackButton from "../../../components/atoms/back-button"
+import Spacer from "../../../components/atoms/spacer"
 import Spinner from "../../../components/atoms/spinner"
 import Tooltip from "../../../components/atoms/tooltip"
+import WidgetContainer from "../../../components/extensions/widget-container"
 import Button from "../../../components/fundamentals/button"
 import DetailsIcon from "../../../components/fundamentals/details-icon"
 import CancelIcon from "../../../components/fundamentals/icons/cancel-icon"
@@ -54,6 +56,7 @@ import useImperativeDialog from "../../../hooks/use-imperative-dialog"
 import useNotification from "../../../hooks/use-notification"
 import useToggleState from "../../../hooks/use-toggle-state"
 import { useFeatureFlag } from "../../../providers/feature-flag-provider"
+import { useWidgets } from "../../../providers/widget-provider"
 import { isoAlpha2Countries } from "../../../utils/countries"
 import { getErrorMessage } from "../../../utils/error-messages"
 import extractCustomerName from "../../../utils/extract-customer-name"
@@ -196,6 +199,8 @@ const OrderDetails = () => {
   useHotkeys("esc", () => navigate("/a/orders"))
   useHotkeys("command+i", handleCopy)
 
+  const { getWidgets } = useWidgets()
+
   const handleDeleteOrder = async () => {
     const shouldDelete = await dialog({
       heading: "Cancel order",
@@ -296,10 +301,22 @@ const OrderDetails = () => {
           </BodyCard>
         ) : (
           <>
+            <div>
+              {getWidgets("order.details.before").map((widget, i) => {
+                return (
+                  <WidgetContainer
+                    key={i}
+                    injectionZone={"order.details.before"}
+                    widget={widget}
+                    entity={order}
+                  />
+                )
+              })}
+            </div>
             <div className="flex space-x-4">
-              <div className="flex h-full w-7/12 flex-col">
+              <div className="gap-y-base flex h-full w-7/12 flex-col">
                 <BodyCard
-                  className={"mb-4 min-h-[200px] w-full"}
+                  className={"min-h-[200px] w-full"}
                   customHeader={
                     <Tooltip side="top" content={"Copy ID"}>
                       <button
@@ -359,7 +376,7 @@ const OrderDetails = () => {
                 <SummaryCard order={order} reservations={reservations || []} />
 
                 <BodyCard
-                  className={"mb-4 h-auto min-h-0 w-full"}
+                  className={"h-auto min-h-0 w-full"}
                   title="Payment"
                   status={
                     <PaymentStatusComponent status={order.payment_status} />
@@ -428,7 +445,7 @@ const OrderDetails = () => {
                   </div>
                 </BodyCard>
                 <BodyCard
-                  className={"mb-4 h-auto min-h-0 w-full"}
+                  className={"h-auto min-h-0 w-full"}
                   title="Fulfillment"
                   status={
                     <FulfillmentStatusComponent
@@ -475,7 +492,7 @@ const OrderDetails = () => {
                   </div>
                 </BodyCard>
                 <BodyCard
-                  className={"mb-4 h-auto min-h-0 w-full"}
+                  className={"h-auto min-h-0 w-full"}
                   title="Customer"
                   actionables={customerActionables}
                 >
@@ -525,9 +542,20 @@ const OrderDetails = () => {
                     </div>
                   </div>
                 </BodyCard>
-                <div className="mt-large">
-                  <RawJSON data={order} title="Raw order" />
+                <div>
+                  {getWidgets("order.details.after").map((widget, i) => {
+                    return (
+                      <WidgetContainer
+                        key={i}
+                        injectionZone={"order.details.after"}
+                        widget={widget}
+                        entity={order}
+                      />
+                    )
+                  })}
                 </div>
+                <RawJSON data={order} title="Raw order" />
+                <Spacer />
               </div>
               <Timeline orderId={order.id} />
             </div>