+++ www/docs/src/components/Badge/index.tsx
@@ -3,14 +3,24 @@ import clsx from "clsx"
 
 export type BadgeProps = {
   className?: string
-  variant: string
+  variant:
+    | "purple"
+    | "purple-dark"
+    | "orange"
+    | "orange-dark"
+    | "green"
+    | "green-dark"
+    | "blue"
+    | "blue-dark"
+    | "red"
+    | "neutral"
 } & React.HTMLAttributes<HTMLSpanElement>
 
 const Badge: React.FC<BadgeProps> = ({ className, variant, children }) => {
   return (
     <span
       className={clsx(
-        "text-compact-x-small-plus py-px px-0.4 rounded-sm border border-solid text-center",
+        "text-compact-x-small-plus px-0.4 rounded-sm border border-solid py-px text-center",
         variant === "purple" &&
           "bg-medusa-tag-purple-bg dark:bg-medusa-tag-purple-bg-dark text-medusa-tag-purple-text dark:text-medusa-tag-purple-text-dark border-medusa-tag-purple-border dark:border-medusa-tag-purple-border-dark",
         variant === "purple-dark" &&
@@ -27,6 +37,10 @@ const Badge: React.FC<BadgeProps> = ({ className, variant, children }) => {
           "bg-medusa-tag-blue-bg dark:bg-medusa-tag-blue-bg-dark text-medusa-tag-blue-text dark:text-medusa-tag-blue-text-dark border-medusa-tag-blue-border dark:border-medusa-tag-blue-border-dark",
         variant === "blue-dark" &&
           "bg-medusa-tag-blue-bg-dark text-medusa-tag-blue-text-dark border-medusa-tag-blue-border-dark",
+        variant === "red" &&
+          "bg-medusa-tag-red-bg dark:bg-medusa-tag-red-bg-dark text-medusa-tag-red-text dark:text-medusa-tag-red-text-dark border-medusa-tag-red-border dark:border-medusa-tag-red-border-dark",
+        variant === "neutral" &&
+          "bg-medusa-tag-neutral-bg dark:bg-medusa-tag-neutral-bg-dark text-medusa-tag-neutral-text dark:text-medusa-tag-neutral-text-dark border-medusa-tag-neutral-border dark:border-medusa-tag-neutral-border-dark",
         "badge",
         className
       )}